Feat/feeed api (#3)
* SAVING for NO PARTICULAR REASON

* find_with_limits done almost. Have to limit docs per id unwanted docs

* find_with_limits DONE and feed too?

* comments, clarity, light, action!

+++ src/config.rs
@@ -10,7 +10,7 @@ pub struct Settings {
     pub db_path: String,
     pub app_name: String,
     pub bakery_trigger_cooldown: i64,
-    pub default_item_per_feed: i32,
+    pub default_item_per_feed: i64,
 }
 
 // #[derive(Debug, Deserialize)]
